+++ glyph-brush/CHANGELOG.md
@@ -28,6 +28,7 @@
   and starts to improve upon it.
 * New _glyph_brush_layout_ now providers `section_index` & `byte_index` for all laid out glyphs. It no longer
   relies on any pixel bounding box info, which isn't fast to query with .otf fonts. This also means invisible glyphs, like ' ', are now generally included and hence will be returned by `GlyphCruncher::glyphs`.
+* Add `GlyphBrush::is_cached` which can be used to tell if a given queued & processed glyph is visible or not.
 * It's faster. **~1.07-1.64x** faster than `0.6` (particularly in the worst case full-layout performance).
   Rasterization is also **~2-7x** faster using _ab_glyph_rasterizer_.